+++ homeassistant/components/trace/const.py
@@ -1,4 +1,4 @@
-"""Shared constants for automation and script tracing and debugging."""
+"""Shared constants for script and automation tracing and debugging."""
 
 DATA_TRACE = "trace"
-STORED_TRACES = 5  # Stored traces per automation
+STORED_TRACES = 5  # Stored traces per script or automation
